requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.10.4
pydantic-settings==2.7.0
python-dotenv==1.0.1

# OCR
pytesseract==0.3.13
Pillow==11.1.0

# PDF
pdf2image==1.17.0
PyMuPDF==1.25.1

# Image preprocessing
opencv-python-headless==4.10.0.84
numpy==2.2.1

# HTTP security
slowapi==0.1.9

# Logging
structlog==24.4.0
